Split out shared glyph plugin game object.
